pyproject.toml
[project]
name = "mailer-back"
version = "0.1.0"
description = "uv로 패키지 관리."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=46.0.1",
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.1.1",
    "asgiref==3.9.1",
    "ruff==0.13.1",
    "setuptools==78.1.1",
    "sqlparse==0.5.3",
    "wheel==0.45.1",
    "annotated-types>=0.7.0",
    "anyio>=4.11.0",
    "google-genai>=1.46.0",
    "google-auth>=2.41.1",
    "graphviz>=0.21",
    "django-extensions>=4.1",
    "pre-commit>=4.3.0",
    "python-jose>=3.5.0",
    "boto3>=1.40.59",
]
